playing around with formatting options
This still needs a lot of work; this is mostlyl for me
to figure out what's possible and how things are likely to work.

The ultimate solution may be to use some sort of templating
system for the output.  But, I'm still not sure how useful
this output fanciness really is.